requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.7.0
APScheduler==3.10.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pipreqs==0.4.13
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pytest==7.4.0
pytest-anyio==0.0.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.4.2
rsa==4.9
shellingham==1.4.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomli==2.0.1
trio==0.22.0
typer==0.7.0
typer-cli==0.0.13
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarg==0.1.9
yarl==1.9.2
